+++ ios/Classes/CountlyFlutterPlugin.m
@@ -253,9 +253,27 @@ FlutterMethodChannel* _channel;
     }else if ([@"getCurrentDeviceId" isEqualToString:call.method]) {
         NSString* deviceId = [Countly.sharedInstance deviceID];
         result(deviceId);
-    }else if ([@"getDeviceIdAuthor" isEqualToString:call.method]) {
-        NSString* deviceIDType = [Countly.sharedInstance deviceIDType];
-        result(deviceIDType);
+    }else if ([@"getDeviceIDType" isEqualToString:call.method]) {
+        CLYDeviceIDType deviceIDType = [Countly.sharedInstance deviceIDType];
+        NSString* deviceIDTypeString = NULL;
+        if([deviceIDType isEqualToString:CLYDeviceIDTypeCustom])
+        {
+            deviceIDTypeString = @"DS";
+        }
+        else if([deviceIDType isEqualToString:CLYDeviceIDTypeIDFV])
+        {
+            deviceIDTypeString = @"SG";
+        }
+        else if([deviceIDType isEqualToString:CLYDeviceIDTypeTemporary])
+        {
+            deviceIDTypeString = @"TID";
+        }
+        else
+        {
+            deviceIDTypeString = @"SG";
+        }
+        result(deviceIDTypeString);
+        
     }else if ([@"changeDeviceId" isEqualToString:call.method]) {
         dispatch_async(dispatch_get_main_queue(), ^ {
         NSString* newDeviceID = [command objectAtIndex:0];